reception fix

+++ inc/Reception.hpp
@@ -29,7 +29,7 @@ private:
 	std::vector<std::string>	strtovec(std::string const &, std::string const &);
 	void	createPizza(std::vector<std::string>);
 
-	size_t	_cookingTime;
+	double	_cookingTime;
 	size_t	_cooksCount;
 	size_t	_resupplyTime;
